+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Search Workspace List
 - Export/Import Workspaces as zip files that can be shared with other people
 - Archive Workspaces to ignore them in the list and show them in a separate list of archived workspaces
+- Better Feedback for successful actions ( "Hitsound diff generated successfully!" notification)
 
 # 0.1.0 & 0.2.0
 - Initial release